requirements.txt
llama-index==0.9.15
langchain==0.1.0
sentence-transformers
weaviate-client
pypdf
python-box
altair<5
streamlit